commit: 카드상품상세보기및 카드상품전체조회

+++ src/main/java/org/cardGGaduekMainService/coupon/couponProduct/domain/CouponProductVO.java
@@ -17,7 +17,7 @@ public class CouponProductVO {
     private Long id;                    // 쿠폰 ID
     private String couponName;          // 쿠폰명
     private LocalDateTime expiredAt;    // 쿠폰 만료일
-    private Long discountValue;         // 쿠폰 할인금액
+    private BigDecimal discountValue;         // 쿠폰 할인금액
     private String couponCategory;      // 쿠폰 카테고리 (FOOD, SHOPPING, MEDICAL, CULTURE, TRANSPORT)
 
 }
